Convert to using torqueclient
The training is changed to use globalview, while the scoring uses a
specific competition.

This change also involves keeping all the data frames in memory instead
of pushing them out to csvs at each step, removing the need for the data
directory.

The training still needs to output a joblib file, but that can be at an
arbitrary location on disk.